Update deps, fix broken install, split AI types
Dependency maintenance:
- Sync node_modules with lockfile (fixes missing @vercel/analytics that
  broke tsc and vite build)
- npm update for same-major bumps (react 19.2.8, i18next-http-backend
  3.0.6, playwright 1.61, prettier 3.9, zustand 5.0.14, etc.)
- npm audit fix (18 -> 8 vulns; remaining are dev-only)
- Add missing @types/node devDependency (fixes lint-tsc-react on Buffer)

Reorganization:
- Remove dead StoreDemo.tsx (imported nowhere)
- Split the 1,160-line types/ai.ts into a types/ai/ directory
  (primitives, commands, events, chat, execution) with a barrel index;
  all import sites resolve unchanged
