+++ src/react/types/ai/index.ts
@@ -1,0 +1,12 @@
+/**
+ * TypeScript types and interfaces for AI integration
+ * Barrel re-exporting the drawing commands, SSE events, chat, and executor
+ * types. Import from "@react/types/ai" to get everything.
+ * Based on the architecture defined in docs/AI.md
+ */
+
+export * from "./primitives";
+export * from "./commands";
+export * from "./events";
+export * from "./chat";
+export * from "./execution";
